Update contact verification & add tooltip component with link AB#17022
