+++ Apps/WebClient/src/ClientApp/src/components/common/InfoTooltipComponent.vue
@@ -6,14 +6,12 @@ interface Props {
     size?: string;
     maxWidth?: number;
     tooltipTestid?: string;
-    contentClass?: string;
 }
 const props = withDefaults(defineProps<Props>(), {
     text: undefined,
     size: "small",
     maxWidth: 300,
     tooltipTestid: undefined,
-    contentClass: undefined,
 });
 
 const openedFromClick = ref(false);
@@ -51,7 +49,6 @@ onBeforeUnmount(() => {
         :model-value="openedFromClick || openedFromHover"
         :open-on-click="false"
         :open-on-hover="false"
-        :content-class="contentClass"
     >
         <template #activator="{ props: slotProps }">
             <v-icon
