+++ Apps/WebClient/src/ClientApp/src/models/notificationPreferenceType.ts
@@ -1,8 +1,14 @@
 import { ProfileNotificationType } from "@/constants/profileNotifications";
 
+export interface TooltipContent {
+    text: string;
+    linkText?: string;
+    href?: string;
+    suffix?: string;
+}
 export interface NotificationPreferenceType {
     id: string;
     type: ProfileNotificationType;
     label: string;
-    tooltip: string;
+    tooltip: TooltipContent;
 }
